Update the structure of source files
* Separate classes from test to allow including
  and reusing different data-structures and algorithms.
