+++ BUILD
@@ -1,9 +1,16 @@
-load("@rules_cc//cc:defs.bzl", "cc_binary")
+load("@rules_cc//cc:defs.bzl", "cc_binary", "cc_library")
+
+cc_library(
+    name = "lib",
+    hdrs = glob(["include/**/*.hpp"]),
+    includes = ["include"],
+)
 
 cc_binary(
     name = "playground",
-    srcs = glob(["src/**/*.cpp"]),
+    srcs = glob(["test/**/*.cpp"]),
     deps = [
+        ":lib",
         "@google_googletest//:gtest_main",
     ],
 )
